Create seek_scan.py
A script that recursively finds all "."  (hidden) files and directories on linux and then uploads them to virustotal automatically using their vt-cli util. You have to have an API key with virustotal in order for it to work but ostensibly can work with any other API. At current the file path for vt is /usr/local/bin/. If you wanted to run it without it uploading to vt just comment out the last line. It'll just create a hidden.txt file in your home dir.